requirements.txt
absl-py==0.9.0
aiohttp==3.5.4
aiohttp-jinja2==1.2.0
aioopenssl==0.5.1
aiosasl==0.4.0
aioxmpp==0.11.0
arrow==0.15.5
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
decorator==4.4.2
dialogflow==0.8.0
dnspython==1.16.0
future==0.18.2
gast==0.2.2
google-api-core==1.16.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
Jinja2==2.11.1
jinja2-time==0.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
multidict==4.5.2
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
protobuf==3.11.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-telegram-bot==12.6.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
singletonify==0.2.3
six==1.14.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
spade==3.1.4
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
timeago==1.0.8
tornado==6.0.4
tzlocal==1.5.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
